requirements.txt
openai>=1.0.0
PyPDF2>=3.0.0
faiss-cpu>=1.7.4
numpy>=1.24.0
streamlit>=1.30.0
tqdm>=4.66.0
python-dotenv>=1.0.0
logging>=0.5.1.2 